refactor: html_parser + models 移入 case_import/，清理 jtn/ 顶层
- html_parser.py → case_import/html_parser.py（唯一消费者是 case_import）
- jtn/models.py 中 JTN 特有模型 → case_import/models.py
- 删除 jtn/models.py（已无外部引用）
- 更新 case_import/ 内所有 import（..models → .models，..html_parser → .html_parser）

jtn/ 顶层现在只剩 adapter.py + __init__.py + auth/ + 各功能模块/
